esp32-webrtc-fastapi-server/requirements.txt
# requirements.txt for ESP32 WebRTC FastAPI Server
fastapi
uvicorn
python-dotenv
loguru
pipecat  # Assuming pipecat is available as a pip package
